refactor: home view moved fo home folder
