+++ src/views/home/home.scss
@@ -1,0 +1,3 @@
+.home {
+    height: 100vh;
+}
